+++ frontend/src/metabase/ui/components/inputs/Select/Select.tsx
@@ -27,7 +27,7 @@ export interface SelectProps<Value extends string | null = string>
   extends Omit<MantineSelectProps, "data" | "onChange" | "value" | "ref"> {
   data: SelectData<Value>;
   value?: Value;
-  onChange?: (newValue: Value, option: ComboboxItem | null) => void;
+  onChange?: (newValue: Value) => void;
 }
 
 function _Select<Value extends string | null>(
